requirements.txt
beautifulsoup4
lxml
requests
pandas
scikit-learn
pdfminer.six
streamlit
